refactor: 재사용 가능 여부에 따른 컴포넌트 디렉토리 재조정
- 재사용 가능한 컴포넌트는 common 폴더 내부로 이동

관련 피드백: https://github.com/woowacourse/react-payments/pull/73#discussion_r862478579
