+++ src/components/Modal.jsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
 
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 import PropTypes from 'prop-types';
