+++ src/stories/Button.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Button from '../components/Button';
+import Button from '../components/common/Button';
 
 export default {
   title: 'Example/Button',
