+++ src/stories/Card.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Card from '../components/Card';
+import Card from '../components/common/Card';
 
 export default {
   title: 'Example/Card',
