+++ src/stories/Input.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Input from '../components/Input';
+import Input from '../components/common/Input';
 
 export default {
   title: 'Example/Input',
