+++ src/stories/mixedComponents.stories.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import Button from '../components/Button';
-import Input from '../components/Input';
+import Button from '../components/common/Button';
+import Input from '../components/common/Input';
 import Modal from '../components/Modal';
 
 export default {
